Individual 14
Fixed an issue where the deadlocked AC gets removed from the aircraft list. This resulted in the main 'for ac in aircraft_list' loop to jump over the next aircraft. Fixed it by instead of deleting the AC, changing its status to "deadlocked". Will also implement this in other planners