Individual 15 + Prioritized 12
Adjusted individual planning algorithm. Everytime self loses the bidding war or has to adjust its path. It will start replanning for all the AC it has already looped over in order to not run into a conflict this way. Also adjusted the deadlock handling in prioritzed to not remove the AC completely but to set it to a "deadlocked" status